+++ dashboard/src/recoil/comments.js
@@ -2,17 +2,11 @@
 import { atom, useRecoilState, useRecoilValue } from 'recoil';
 import { currentTeamState, organisationState, userState } from '../recoil/auth';
 import useApi from '../services/api';
-import { getData, useStorage } from '../services/dataManagement';
 import { capture } from '../services/sentry';
 
 export const commentsState = atom({
   key: 'commentsState',
   default: [],
-});
-
-export const commentsLoadingState = atom({
-  key: 'commentsLoadingState',
-  default: true,
 });
 
 export const useComments = () => {
@@ -22,37 +16,6 @@ export const useComments = () => {
   const API = useApi();
 
   const [comments, setComments] = useRecoilState(commentsState);
-  const [lastRefresh, setLastRefresh] = useStorage('last-refresh-comments', 0);
-  const [loading, setLoading] = useRecoilState(commentsLoadingState);
-
-  const setCommentsFullState = (newComments) => {
-    if (newComments) setComments(newComments);
-    setLoading(false);
-    setLastRefresh(Date.now());
-  };
-  const setBatchData = (newActions) => setComments((actions) => [...actions, ...newActions]);
-
-  const refreshComments = async (setProgress, initialLoad) => {
-    setLoading(true);
-    try {
-      setCommentsFullState(
-        await getData({
-          collectionName: 'comment',
-          data: comments,
-          isInitialization: initialLoad,
-          setProgress,
-          setBatchData,
-          lastRefresh,
-          API,
-        })
-      );
-      return true;
-    } catch (e) {
-      setLoading(false);
-      capture(e.message, { extra: { response: e.response } });
-      return false;
-    }
-  };
 
   const deleteComment = async (id) => {
     const res = await API.delete({ path: `/comment/${id}` });
@@ -95,9 +58,7 @@ export const useComments = () => {
   };
 
   return {
-    loading,
     comments,
-    refreshComments,
     deleteComment,
     addComment,
     updateComment,