+++ dashboard/src/recoil/territoryObservations.js
@@ -1,17 +1,11 @@
 import { organisationState } from './auth';
 import useApi from '../services/api';
-import { getData, useStorage } from '../services/dataManagement';
 import { capture } from '../services/sentry';
 import { atom, selector, useRecoilState, useRecoilValue } from 'recoil';
 
 export const territoryObservationsState = atom({
   key: 'territoryObservationsState',
   default: [],
-});
-
-export const territoriesObservationsLoadingState = atom({
-  key: 'territoriesObservationsLoadingState',
-  default: true,
 });
 
 export const customFieldsObsSelector = selector({
@@ -25,43 +19,10 @@ export const customFieldsObsSelector = selector({
 
 export const useTerritoryObservations = () => {
   const [territoryObservations, setTerritoryObs] = useRecoilState(territoryObservationsState);
-  const [loading, setLoading] = useRecoilState(territoriesObservationsLoadingState);
-  const [lastRefresh, setLastRefresh] = useStorage('last-refresh-observations', 0);
 
   const API = useApi();
 
   const customFieldsObs = useRecoilValue(customFieldsObsSelector);
-
-  const setTerritoryObsFullState = (newTerritoryObservations) => {
-    if (newTerritoryObservations) {
-      setTerritoryObs(newTerritoryObservations);
-      setLoading(false);
-    }
-    setLastRefresh(Date.now());
-  };
-
-  const setBatchData = (newObs) => setTerritoryObs((territoryObservations) => [...territoryObservations, ...newObs]);
-
-  const refreshTerritoryObs = async (setProgress) => {
-    setLoading(true);
-    try {
-      const data = await getData({
-        collectionName: 'territory-observation',
-        data: territoryObservations,
-        isInitialization: true,
-        setProgress,
-        setBatchData,
-        lastRefresh,
-        API,
-      });
-      setTerritoryObsFullState(data);
-      return true;
-    } catch (e) {
-      capture(e.message, { extra: { response: e.response } });
-      setLoading(false);
-      return false;
-    }
-  };
 
   const deleteTerritoryObs = async (id) => {
     const res = await API.delete({ path: `/territory-observation/${id}` });
@@ -104,9 +65,7 @@ export const useTerritoryObservations = () => {
 
   return {
     territoryObservations,
-    loading,
     customFieldsObs,
-    refreshTerritoryObs,
     setTerritoryObs,
     deleteTerritoryObs,
     addTerritoryObs,